Fix four compile errors from CI build
- main.dart: navigationBarColor → systemNavigationBarColor (correct SystemUiOverlayStyle field)
- notification_service.dart: add required uiLocalNotificationDateInterpretation arg to both zonedSchedule calls
- api_service.dart: explicit <int, FFData>{} types on early-return records to fix Map<dynamic,dynamic> inference
- settings_screen.dart: s.minPrice.toDouble() to match _NumberField's double parameter

+++ bounty-hunter/lib/screens/settings_screen.dart
@@ -159,7 +159,7 @@ class _SettingsScreenState extends State<SettingsScreen> {
             children: [
               _NumberField(
                 label: 'Min reward (\$)',
-                value: s.minPrice,
+                value: s.minPrice.toDouble(),
                 min: 0,
                 step: 100000,
                 onChanged: (v) => _updateSettings(s.copyWith(minPrice: v.round())),
